moved ncore_ from molbe.helper to shared.helper
