+++ src/quemb/molbe/ube.py
@@ -19,7 +19,7 @@ import numpy
 from pyscf import ao2mo
 
 from quemb.molbe.be_parallel import be_func_parallel_u
-from quemb.molbe.helper import unused
+from quemb.shared.helper import unused
 from quemb.molbe.mbe import BE
 from quemb.molbe.pfrag import Frags
 from quemb.molbe.solver import be_func_u
