+++ src/quemb/shared/helper.py
@@ -1,4 +1,5 @@
-from typing import Callable, TypeVar
+import sys
+from typing import Any, Callable, TypeVar
 
 Function = TypeVar("Function", bound=Callable)
 
@@ -33,3 +34,30 @@ def add_docstring(doc: str) -> Callable[[Function], Function]:
         return f
 
     return update_doc
+
+
+def unused(*args: Any) -> None:
+    for arg in args:
+        del arg
+
+
+def ncore_(z: int) -> int:
+    if 1 <= z <= 2:
+        nc = 0
+    elif 2 <= z <= 5:
+        nc = 1
+    elif 5 <= z <= 12:
+        nc = 1
+    elif 12 <= z <= 30:
+        nc = 5
+    elif 31 <= z <= 38:
+        nc = 9
+    elif 39 <= z <= 48:
+        nc = 14
+    elif 49 <= z <= 56:
+        nc = 18
+    else:
+        print("Ncore not computed in helper.ncore(), add it yourself!", flush=True)
+        print("exiting", flush=True)
+        sys.exit()
+    return nc
